+++ ft_memmove.c
@@ -6,13 +6,15 @@
 /*   By: bhildebr <bhildebr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/26 16:33:35 by bhildebr          #+#    #+#             */
-/*   Updated: 2023/08/12 15:25:37 by bhildebr         ###   ########.fr       */
+/*   Updated: 2023/08/12 16:30:04 by bhildebr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memmove(void *dest, const void *src, unsigned long n)
+#include "libft.h"
+
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
-	unsigned long	index;
+	size_t	index;
 
 	if (dest == ((void *)0) && src == ((void *)0))
 		return ((void *)0);
