+++ ft_strchr.c
@@ -6,16 +6,15 @@
 /*   By: bhildebr <bhildebr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/27 11:30:21 by bhildebr          #+#    #+#             */
-/*   Updated: 2023/08/11 17:05:11 by bhildebr         ###   ########.fr       */
+/*   Updated: 2023/08/12 18:06:04 by bhildebr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-// The strchr() function returns a pointer to the first occurrence of the 
-// character c in the string s.
+#include "libft.h"
 
 char	*ft_strchr(const char *s, int c)
 {
-	unsigned long	index;
+	size_t	index;
 
 	index = 0;
 	while (s[index])
