+++ ft_strmapi.c
@@ -12,12 +12,10 @@
 
 #include "libft.h"
 
-#include <stdlib.h>
-
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	unsigned long	i;
-	char			*str;
+	size_t	i;
+	char	*str;
 
 	i = 0;
 	str = malloc((ft_strlen(s) + 1) * sizeof(char));
