+++ ft_strncmp.c
@@ -6,21 +6,15 @@
 /*   By: bhildebr <bhildebr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/27 11:59:40 by bhildebr          #+#    #+#             */
-/*   Updated: 2023/08/11 17:17:56 by bhildebr         ###   ########.fr       */
+/*   Updated: 2023/08/12 18:06:29 by bhildebr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-// The strcmp() function compares the two strings s1 and s2.  The locale is not
-// taken into account (for a locale-aware comparison, see strcoll(3)).  It 
-// returns an integer less than, equal to, or greater than zero if s1 is found,
-// respectively, to be less than, to match, or be greater than s2.
+#include "libft.h"
 
-// The strncmp() function is similar, except it compares only the first 
-// (at most) n bytes of s1 and s2.
-
-int	ft_strncmp(const char *s1, const char *s2, unsigned long n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned long	index;
+	size_t	index;
 
 	index = 0;
 	if (s1 == 0 && s2 == 0)
