+++ ft_strnstr.c
@@ -6,22 +6,16 @@
 /*   By: bhildebr <bhildebr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/27 13:06:58 by bhildebr          #+#    #+#             */
-/*   Updated: 2023/08/11 17:19:39 by bhildebr         ###   ########.fr       */
+/*   Updated: 2023/08/12 18:06:34 by bhildebr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-// The strnstr() function locates the first occurrence of the null-terminated 
-// string little in the string big, where not more than len characters are 
-// searched.  Characters that appear after a ‘\0’ character are not searched.  
-// Since the strnstr() function is a FreeBSD specific API, it should only be 
-// used when portability is not a concern.
-
 #include "libft.h"
 
-char	*ft_strnstr(const char *big, const char *little, unsigned long len)
+char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
-	unsigned long	big_index;
-	unsigned long	little_index;
+	size_t	big_index;
+	size_t	little_index;
 
 	if (little[0] == '\0')
 		return ((char *)big);
